parlant/requirements.txt
parlant[gemini]>=3.0.0
google-genai>=0.2.0
google-api-core>=2.11.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
cachetools>=5.3.0
pytest
pytest-asyncio
pytest-httpx
hypothesis>=6.0.0

# Async webhook orchestration dependencies
hatchet-sdk>=0.25.0
redis>=5.0.0
redis[hiredis]>=5.0.0
prometheus-client>=0.19.0
asyncio-mqtt>=0.16.0
psutil>=5.9.0
